Moved findJob and test to db files; Stubbed out create_job route;
decided that finder methods should really belong in the db. the create functions I'm not so sure about though,
because I feel like the encryption functions should be private and in the arylib file, but the createUser function requires those methods so...
I just left them in arylib for now.